third_party/WebKit/LayoutTests/accessibility/aria-disabled-expected.txt

 
link  submit
treeitem
option
tab
gridcell
checkbox
This tests that the aria-disabled attribute works. The text field should not be enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textField.restrictions is "disabled"
PASS textField.restrictions is "none"
PASS group.restrictions is "disabled"
PASS link.restrictions is "disabled"
PASS button.restrictions is "none"
PASS treeitem.restrictions is "disabled"
PASS option.restrictions is "disabled"
PASS tab.restrictions is "disabled"
PASS gridcell.restrictions is "disabled"
PASS checkbox.restrictions is "disabled"
PASS successfullyParsed is true

TEST COMPLETE
